+++ NEWS.md
@@ -26,6 +26,10 @@
 -   Small Vignettes and documentation improvements
 -   Proper testing of examples
 -   `delta` parameter of `gips()` has to be bigger than `1`. We used to restrict it as bigger than `2`.
+-   `project_matrix()` will show a warning when the non-positive-semi-definite matrix is passed as an `S` argument.
+-   `project_matrix()` will preserve colnames and rownames of a matrix.
+-   `D_matrix` will be checked for containing any `NaN` or `Inf` values.
+
 
 # gips 1.0.0